Split metrics test between Open and EE
